amdgpuadapterinfo
